beta.polyclass {polspline}	R Documentation

Polyclass: polychotomous regression and multiple classification

Description

Produces a beta-plot for a polyclass object.

Usage

beta.polyclass(fit, which, xsp = 0.4, cex) 

Arguments

fit
polyclass object, typically the result of polyclass.

which
which classes should be compared? Default is to compare all classes.

xsp
location of the vertical line to the left of the axis. Useful for making high quality, device dependent, graphics.

cex
character size. Default is whatever the present character size is. Useful for making high quality, device dependent, graphics.

Value

A beta plot. One line for each basis function. The left part of the plot indicates the basis function, the right half the relative location of the betas (coefficients) of that basis function, normalized with respect to parent basis functions, for all classes. The scaling is supposed to suggest a relative importance of the basis functions. This may suggest which basis functions are important for separating particular classes.

Note

This is not a generic function, and the complete name, beta.polyclass, has to be specified.

Author(s)

Charles Kooperberg clk@fredhutch.org.

References

Charles Kooperberg, Smarajit Bose, and Charles J. Stone (1997). Polychotomous regression. Journal of the American Statistical Association, 92, 117–127.

Charles J. Stone, Mark Hansen, Charles Kooperberg, and Young K. Truong. The use of polynomial splines and their tensor products in extended linear modeling (with discussion) (1997). Annals of Statistics, 25, 1371–1470.

See Also

polyclass, plot.polyclass, summary.polyclass, cpolyclass, ppolyclass, rpolyclass.

Examples

data(iris)
fit.iris <- polyclass(iris[,5], iris[,1:4])
beta.polyclass(fit.iris)

[Package polspline version 1.1.20 Index]