clspec {polspline}	R Documentation

Lspec: logspline estimation of a spectral distribution

Description

Autocorrelations, autocovariances (clspec), spectral densities and line spectrum (dlspec), spectral distributions (plspec) or a random time series(rlspec) from a model fitted with lspec.

Usage

clspec(lag, fit, cov = TRUE, mm) 
dlspec(freq, fit) 
plspec(freq, fit, mm) 
rlspec(n, fit, mean = 0, cosmodel = FALSE, mm)

Arguments

lag
vector of integer-valued lags for which the autocorrelations or autocorrelations are to be computed.

fit
lspec object, typically the result of lspec.

cov
compute autocovariances (TRUE) or autocorrelations (FALSE).

mm
number of points used in integration and the fft. Default is the smallest power of two larger than max(fit\$sample, max(lag),1024) for clspec and plspec or the smallest power of two larger than max(fit\$sample, n, max(lag), 1024) for (rlspec).

freq
vector of frequencies. For plspec frequencies should be between -pi and pi.

n
length of the random time series to be generated.

mean
mean level of the time series to be generated.

cosmodel
indicate that the data should be generated from a model with constant harmonic terms rather than a true Gaussian time series.

Value

Autocovariances or autocorrelations (clspec); values of the spectral distribution at the requested frequencies. (plspec); random time series of length n (rlspec); or a list with three components (dlspec):

d
the spectral density evaluated at the vector of frequencies,

modfreq
modified frequencies of the form (2pi*j/T) that are close to the frequencies that were requested,

m
mass of the line spectrum at the modified frequencies.

Author(s)

Charles Kooperberg clk@fredhutch.org.

References

Charles Kooperberg, Charles J. Stone, and Young K. Truong (1995). Logspline Estimation of a Possibly Mixed Spectral Distribution. Journal of Time Series Analysis, 16, 359-388.

Charles J. Stone, Mark Hansen, Charles Kooperberg, and Young K. Truong. The use of polynomial splines and their tensor products in extended linear modeling (with discussion) (1997). Annals of Statistics, 25, 1371–1470.

See Also

lspec, plot.lspec, summary.lspec.

Examples

data(co2)
co2.detrend <- lm(co2~c(1:length(co2)))$residuals
fit <- lspec(co2.detrend)
clspec(0:12,fit)
plspec((0:314)/100, fit)
dlspec((0:314)/100, fit)
rlspec(length(co2),fit)

[Package polspline version 1.1.20 Index]