persp.polymars {polspline}	R Documentation

Polymars: multivariate adaptive polynomial spline regression

Description

This function is not intended for direct use. It is called by plot.polymars.

Usage

## S3 method for class 'polymars'
persp(x, predictor1, predictor2, response, n = 33,
xlim, ylim, xx, contour.polymars, main, intercept, ...) 

Arguments

x, predictor1, predictor2
this function is not intended to be called directly.

response, n, xlim, ylim
this function is not intended to be called directly.

xx, contour.polymars
this function is not intended to be called directly.

main, intercept, ...
this function is not intended to be called directly.

Details

This function produces a 3-d contour or perspective plot. It is intended to be called by plot.polymars.

Author(s)

Martin O'Connor.

References

Charles Kooperberg, Smarajit Bose, and Charles J. Stone (1997). Polychotomous regression. Journal of the American Statistical Association, 92, 117–127.

Charles J. Stone, Mark Hansen, Charles Kooperberg, and Young K. Truong. The use of polynomial splines and their tensor products in extended linear modeling (with discussion) (1997). Annals of Statistics, 25, 1371–1470.

See Also

polymars, plot.polymars.

[Package polspline version 1.1.20 Index]