summary.lspec {polspline}	R Documentation

Lspec: logspline estimation of a spectral distribution

Description

Summary of a model fitted with lspec

Usage

## S3 method for class 'lspec'
summary(object, ...) 
## S3 method for class 'lspec'
print(x, ...) 

Arguments

object,x
lspec object, typically the result of lspec.

...
other options are ignored.

Details

These function produce an identical printed summary of an lspec object.

Author(s)

Charles Kooperberg clk@fredhutch.org.

References

Charles Kooperberg, Charles J. Stone, and Young K. Truong (1995). Logspline Estimation of a Possibly Mixed Spectral Distribution. Journal of Time Series Analysis, 16, 359-388.

Charles J. Stone, Mark Hansen, Charles Kooperberg, and Young K. Truong. The use of polynomial splines and their tensor products in extended linear modeling (with discussion) (1997). Annals of Statistics, 25, 1371–1470.

See Also

lspec, plot.lspec, clspec, dlspec, plspec, rlspec.

Examples

data(co2)
co2.detrend <- lm(co2~c(1:length(co2)))$residuals
fit <- lspec(co2.detrend)
summary(fit)

[Package polspline version 1.1.20 Index]