unstrip {polspline}	R Documentation

Reformat data as vector or matrix

Description

This function tries to convert a date.frame or a matrix to a no-frills matrix without labels, and a vector or time-series to a no-frills vector without labels.

Usage

unstrip(x) 

Arguments

x
one- or two-dimensional object.

Details

Many of the functions for logspline, oldlogspline, lspec, polyclass, hare, heft, and polymars were written in the “before data.frame” era; unstrip attempts to keep all these functions useful with more advanced input objects. In particular, many of these functions call unstrip before doing anything else.

Value

If x is two-dimensional a matrix without names, if x is one-dimensional a numerical vector

Author(s)

Charles Kooperberg clk@fredhutch.org.

Examples

data(co2)
unstrip(co2)
data(iris)
unstrip(iris)

[Package polspline version 1.1.20 Index]